feat: se mueve funcion GetJefeDependencia a su propio fichero
